refactor(listen): split top bar search and user menu
